genetictest/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect2==0.8.0
Django==5.2.9
djangorestframework==3.17.1
freetype-py==2.5.1
html5lib==1.1
idna==3.11
lxml==6.0.2
oscrypto==1.3.0
pillow==12.0.0
pycairo==1.29.0
pycparser==2.23
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.4.2
python-bidi==0.6.7
PyYAML==6.0.3
reportlab==4.4.6
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.4
svglib==1.6.0
tinycss2==1.5.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.6.2
webencodings==0.5.1
xhtml2pdf==0.2.17
